Fixed auth plugin
The auth plugin was broken due to an incorrect table name that should
have been a config variable ($this->config['table']). I fixed it, and
also renamed some functions.

The checkCookie function was moved into the construct, as it was only
called once, and there were a couple other miscellaneous cleanups.