Fixed issue one
The plugin doesn't have to be loaded by the controller for get_plugin to
work anymore: if a plugin calls get_plugin and the plugin isn't loaded,
it loads it into the controller and then sends a reference to the
plugin.

Also removed the exists method - it was somewhat pointless.