Added a debug mode
I've added a debug mode. It can be defined in /config.php - simply
'debug' (set to true by default). Currently it doesn't do much - it just
turns error reporting on.

Also fixed in a minor bug in the config file where something was being
defined as (string) 'true', instead of the boolean equivilant.