Created the "home" controller and "home_body" view
Started transforming it into a social network. Firstly deleted the old
calc and hello controllers and the various calc views from the original
framework, then created a new controller and view. As issue 1 has not
yet been resolved, there are a tonne of $this->load()s that don't seem
to do anything. Hopefully they will be removed when issue 1 is fixed.