Added global DocBlock stuff, and changed the if (!IN_LYNX)
The global DocBlock stuff is the standard stuff - package name, ID
thing, copyright and a link to the creative commons license that the
lynx framework is released under.

Also, the if (!IN_LYNX)'s have been changed to if (!defined('IN_LYNX'));
this will produce nicer errors.

Added namespaces to the hash class, and an index.html - looks like it
missed out when I was introducing them.