Started developing a feed plugin.
This plugin will be one of the most important things about the
framework, so I'm going to do a good job of it! So far I have only added
a basic post and get system, will add more later. See DocBlocks for more
info (I'm too lazy to type it all here)